--- api/src/opentrons/drivers/asyncio/thermocycler/__init__.py
@@ -1,10 +1,0 @@
-from .driver import ThermocyclerDriver
-from .simulator import SimulatingDriver
-from .abstract import AbstractThermocyclerDriver
-
-
-__all__ = [
-    "ThermocyclerDriver",
-    "SimulatingDriver",
-    "AbstractThermocyclerDriver"
-]
