+++ api/src/opentrons/drivers/thermocycler/__init__.py
@@ -1,7 +1,10 @@
-from .driver import Thermocycler, SimulatingDriver, ThermocyclerError
+from .driver import ThermocyclerDriver
+from .simulator import SimulatingDriver
+from .abstract import AbstractThermocyclerDriver
+
 
 __all__ = [
-    'Thermocycler',
-    'SimulatingDriver',
-    'ThermocyclerError'
+    "ThermocyclerDriver",
+    "SimulatingDriver",
+    "AbstractThermocyclerDriver"
 ]
