+++ api/tests/opentrons/drivers/thermocycler/test_driver.py
@@ -4,7 +4,7 @@ import pytest
 from mock import AsyncMock
 from opentrons.drivers.asyncio.communication.serial_connection import \
     SerialConnection
-from opentrons.drivers.asyncio.thermocycler import driver
+from opentrons.drivers.thermocycler import driver
 from opentrons.drivers.command_builder import CommandBuilder
 from opentrons.drivers.types import (
     Temperature, PlateTemperature, ThermocyclerLidStatus
